requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfminer.six==20221105
transformers==4.35.2
torch>=2.7.1
numpy==1.24.3
scikit-learn==1.3.2
python-dotenv==1.0.0 